IR: move graph relative code from IRA to IR
